added some tests and finished implementing features on last commit
I'm commiting now to start changing the server killing feature to a
socket instead of a channel, so that I can kill it from another process
and not only from another thread
