+++ wallets/receiver/public.pem
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAknDvqh7nstvp4ig+Qp4K
+0Dp3G3nVQ20b093rJS4a37CaHaMX3qnlU+Mk0zWoHD3WFisUaypAHu7UFtfKp6oH
+Iw2Ns7Cl7NLlMDQA6nOHbRaoYjpzpG5EjUD9w27VfO6AJebCTDKnumSxfAjXPVCh
+FzpC3X7Ve81CVILJSuQVtV8VQ7rwelfNU3WToChMQ1LAE3YQJGgj8O/SB+FgXPoV
+9oeYakhPZVADk1czmF+WCaPmKASQYMO53DRwAcnoqMM6QZO16NLIqHgLBxbKLLNz
+wsvNWTR3Duu4ZUl7ZI+vosMqWkBNjGC+xdJiHNReuc2xSzyprAXNxcUTLniy30gP
+bwIDAQAB
+-----END PUBLIC KEY-----
