+++ wallets/sender/public.pem
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAr0+kMrWEaHrNlgU/KxPg
+zDrtiVEqj5yXVTkCFeVCQ2UAIjyMuxi9EtazDryZS6EWun4NU/Lb0/8OsUcThAhN
+gLmqTz2B6MYzNA2hy/oeQI/7Nsag8r20mR7BDlQHGZcE5kUmsKH5aC22mQ6vVNbi
+JZ7lDIYDuv0+TEJoAgK6c0cbeWYWe6Qpy+xgICI57KaoRywwCk8XoS/afeS94EiD
+jAG31PKC7AqLotKcJXmwpl8CzgcktJ2HzUPLyMMI+kli2jBnth2wE23P1QGEgEBv
+r9EJASNjCS/gDwxuEAeRng8vCR60byLQ81gzISlCClc/yXxf4/b7iA1s+e2ooBV1
+nQIDAQAB
+-----END PUBLIC KEY-----
